requirements.txt
google-cloud-core~=2.4.1
google-cloud-documentai~=2.21.1
google-cloud-storage~=2.14.0
google-cloud-documentai-toolbox~=0.12.0a0
simplejson~=3.19.2
python-dotenv~=1.0.1
hocr-tools~=1.1.1
Pillow~=10.2.0
pathlib
PyMuPDF~=1.23.19
reportlab~=4.0.9
python-magic~=0.4.27
virtualenv~=20.25.0